Update Mule Resource attributes (#53)
* Update Mule Resource attributes

. Removes the mule.env attribute. This is a user-defined attribute and
not present on CH.
. Introduce mule.organization.id which is more mule- user friendly name.
Value is same as CS Organization Id defined by CloudHub, unless
`organization.id` system property is defined.

* Removing the usage of organization.id sys prop

CloudHub seems to have that system property and it's value isn't what we
assumed to be. Removing use of this system property for organization id.
We will use CS Organization id for this user friendly attribute name.